test: merged tests file
